separated html extensions
